+++ Cargo.toml
@@ -166,7 +166,21 @@ wasm-bindgen-futures = "0.4.50"
 js-sys = "0.3"
 serde-wasm-bindgen = "0.6"
 serde_bytes = "0.11"
-web-sys = { version = "0.3", features = ["Performance", "Window", "Document", "Element", "HtmlCanvasElement", "Window"] }
+web-sys = { version = "0.3", features = [
+    "Performance", 
+    "Window", 
+    "Document", 
+    "Element", 
+    "HtmlCanvasElement", 
+    "DedicatedWorkerGlobalScope",
+    "CacheStorage",
+    "Cache",
+    "Request",
+    "Response",
+    "ResponseInit", 
+    "Headers",    
+    "Url",
+] }
 web-time = "1.1"
 copypwasmta = { git = "https://github.com/robtfm/copypwasmta" }
 web-fs = "0.2.6"
