+++ crates/image_processing/Cargo.toml
@@ -41,3 +41,7 @@ gltf-json = "1.4"
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 web-fs = { workspace = true }
 futures-lite = { workspace = true }
+web-sys = { workspace = true }
+wasm-bindgen = { workspace = true }
+wasm-bindgen-futures = { workspace = true }
+js-sys = { workspace = true }
